[Refactor][Zeta] Group savepoint feature under checkpoint.savepoint
Move the engine-side savepoint feature into a dedicated package group:
- checkpoint.savepoint.SavepointSaveSession (savepoint write session)
- checkpoint.savepoint.serialization.* (engine-wire v1 DTOs, CheckpointWireCodec,
  LegacyCheckpointReader, SavepointReader/ReaderV1/Registry,
  SavepointIncompatibleException)
- tests follow the same grouping (wire compatibility, reader registry,
  restore selection)

Runtime checkpoint classes stay under checkpoint.*; storage-facing DTOs stay
under checkpoint.storage.savepoint.*.
