+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/checkpoint/savepoint/serialization/WireActionState.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.engine.server.checkpoint.serialization;
+package org.apache.seatunnel.engine.server.checkpoint.savepoint.serialization;
 
 import io.protostuff.Tag;
 import lombok.AllArgsConstructor;
